replace deprecated np.int > np.int32, np.float > np.float32, code cleaning

+++ data/mask_cspad_2x1_edges.py
@@ -1,21 +1,16 @@
 #!/usr/bin/env python
-#--------------------
 
-from __future__ import print_function
-from __future__ import division
 import os
 import sys
 import numpy as np
-    
-#--------------------
 
 def mask_cspad_2x1_edges() :
     """Returns CSPAD array of shape (5920, 388) with mask of 2x1 edges
     """
     segs, rows, cols, colsh = 32, 185, 388, 388//2
-    zero_col = np.zeros(rows,dtype=np.int)
-    zero_row = np.zeros(cols,dtype=np.int)
-    mask2x1  = np.ones((rows,cols),dtype=np.int)
+    zero_col = np.zeros(rows,dtype=np.int32)
+    zero_row = np.zeros(cols,dtype=np.int32)
+    mask2x1  = np.ones((rows,cols),dtype=np.int32)
     mask2x1[0, :]      = zero_row # mask top    edge
     mask2x1[-1,:]      = zero_row # mask bottom edge
     mask2x1[:, 0]      = zero_col # mask left   edge
@@ -23,10 +18,9 @@ def mask_cspad_2x1_edges() :
     mask2x1[:,colsh-1] = zero_col # mask central-left  column
     mask2x1[:,colsh]   = zero_col # mask central-right column
     #print 'mask2x1:\n', mask2x1
-    
+
     return np.vstack([mask2x1 for seg in range(segs)])
 
-#--------------------
 
 if __name__ == '__main__' :
 
@@ -42,4 +36,4 @@ if __name__ == '__main__' :
 
     sys.exit('The End')
 
-#--------------------
+# EOF
